pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twitter-scraper-pipeline"
version = "1.0.0"
description = "Multi-engine Twitter scraping pipeline with twscrape and twikit"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Searching",
]

dependencies = [
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "loguru>=0.7.0",
    "orjson>=3.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "twscrape>=0.17.0",
    "fake-useragent>=1.4.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
twikit = [
    "twikit>=2.3.0",
    "pyotp>=2.9.0",
    "filetype>=1.2.0",
    "webvtt-py>=0.5.0",
    "m3u8>=0.3.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
twitter-pipeline = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "engines*", "core*", "models*", "output*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false